refactor: move get formatted time to utils/date
