+++ site/src/components/timetables/single_timetable_row/index.tsx
@@ -1,6 +1,6 @@
 import type { Locale } from '@typings/common'
 
-import { getFormattedTime } from '@utils/get_formatted_time'
+import { getFormattedTime } from '@utils/date'
 
 import {
   StyledCancelled,
